+++ linea.h
@@ -27,13 +27,13 @@ int obtener_cantidad_productos_linea(linea lin);
 int obtener_precio_unitario_linea(linea lin);
 
 // devuelve la multiplicación entre precio unitario y cantidad
-int importe_linea(linea lin);
+float importe_linea(linea lin);
 
 // devuelve la suma entre importe_linea y iva_linea
-int importe_total_linea(linea lin);
+float importe_total_linea(linea lin);
 
 // devuelve la multiplicación entre importe y IVA
-int iva_linea(linea lin);
+float iva_linea(linea lin);
 
 #endif // LINEA_H_INCLUDED
